+++ src/core/engine/index.ts
@@ -6,4 +6,5 @@
 
 export * from "./evaluate";
 export * from "./magic";
+export * from "./recruitment";
 export * from "./munitions";